requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
Django==3.2.17
django-allauth==0.52.0
django-filter==22.1
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
